+++ CompanionPackages~/com.playserv.schema-tool/CHANGELOG.md
@@ -2,6 +2,11 @@
 
 ## [Unreleased]
 
+## [0.3.7] - 2026-08-12
+
+- Aligned the external tool, package, and core SDK dependency versions with
+  `0.3.7`.
+
 ## [0.3.6] - 2026-08-12
 
 - Aligned the external tool, package, and core SDK dependency versions with